Added Numeric Mutation
Numeric mutation (Evett & Fdez, 1998) is a genetic operaiton that modifies all scalar constants in a GP tree.

This update to TurboGP includes the prototypical method in GPOperators modules, as well as a wrapper for SimpleRegresor class.

TODO:

It still remains to add similar wrappers to BinaryClassifier and other individual classes.

Notice however, that in its current form in SimpleRegresor only supports minimization problems; therefore precautions should be taken when deployed for binary classifiers or other type of problems where fitness function is to be maximized.