Fixed Keijzer7, added 2d and 3d plotting functions, and added new test functions.
Keijzer7 had a bug because the sumation began in 0, has now been fixed.

Added 2D and 3D plotting function for easy observation of artificial regression probles;
mainly to determine ideal -or, difficulty level- of a given function according to the range used.

Also added a new collection of test functions, linobi's, my own suggestions; these are particularly
difficult functions for GP to approximate, as long as trigonometric function are not part of the
primitives set.